Added support for publicUserDocumentationUrl

+++ run_tests.sh
@@ -1,5 +1,11 @@
 #!/bin/sh
 
-pipenv install --dev 
+EXIST=`command -v pipenv`
+
+if [ -z $EXIST ]; then
+    pipenv install --dev 
+else 
+    echo "pipenv install --dev already run"; 
+fi
 
 pipenv run green -vv --run-coverage --failfast "test" "$@"
